requirements.txt
flask
requests
schedule
python-dotenv
PyGithub